feat: added more fields to e2e tests for email mode form creation
