+++ __tests__/e2e/constants/field.ts
@@ -1,0 +1,297 @@
+// /**
+//  * Creates a field for e2e tests with default options (visible, required, not blank).
+//  * @param {Object} fieldObj Custom options of the field.
+//  * @param {string} fieldObj.title Mandatory title of field
+//  * @param {string} fieldObj.fieldType Type of field. Mandatory unless makeField is
+//  * being called to create an artificial field for verified SingPass/CorpPass data.
+//  * @param {string} fieldObj.val Mandatory answer to field
+//  */
+// function makeField(fieldObj) {
+//   return Object.assign(
+//     {
+//       required: true,
+//       isVisible: true,
+//       isLeftBlank: false,
+//     },
+//     fieldObj,
+//   )
+// }
+
+import { format } from 'date-fns'
+import { keyBy } from 'lodash'
+import {
+  AttachmentFieldBase,
+  AttachmentSize,
+  BasicField,
+  CheckboxFieldBase,
+  DateFieldBase,
+  DecimalFieldBase,
+  DropdownFieldBase,
+  EmailFieldBase,
+  FieldBase,
+  HomenoFieldBase,
+  ImageFieldBase,
+  LongTextFieldBase,
+  MobileFieldBase,
+  NricFieldBase,
+  NumberFieldBase,
+  RadioFieldBase,
+  RatingFieldBase,
+  RatingShape,
+  SectionFieldBase,
+  ShortTextFieldBase,
+  StatementFieldBase,
+  TableFieldBase,
+  UenFieldBase,
+  YesNoFieldBase,
+} from 'shared/types'
+
+type E2eFieldSingleValue = { val: string }
+type E2eFieldMultiValue = { val: string[] }
+type E2eFieldTableValue = { val: string[][] }
+type E2eFieldFilepath = { path: string }
+
+type E2ePickFieldMetadata<T extends FieldBase, K extends keyof T> = Pick<
+  T,
+  'title' | 'fieldType' | K
+> &
+  Partial<Pick<T, 'description' | 'required'>>
+
+export type E2eFieldMetadata =
+  | (E2ePickFieldMetadata<AttachmentFieldBase, 'attachmentSize'> &
+      E2eFieldSingleValue &
+      E2eFieldFilepath)
+  | (E2ePickFieldMetadata<
+      CheckboxFieldBase,
+      | 'fieldOptions'
+      | 'othersRadioButton'
+      | 'ValidationOptions'
+      | 'validateByValue'
+    > &
+      E2eFieldMultiValue)
+  | (E2ePickFieldMetadata<DateFieldBase, 'dateValidation'> &
+      E2eFieldSingleValue)
+  | (E2ePickFieldMetadata<
+      DecimalFieldBase,
+      'ValidationOptions' | 'validateByValue'
+    > &
+      E2eFieldSingleValue)
+  | (E2ePickFieldMetadata<DropdownFieldBase, 'fieldOptions'> &
+      E2eFieldSingleValue)
+  | (E2ePickFieldMetadata<
+      EmailFieldBase,
+      | 'isVerifiable'
+      | 'autoReplyOptions'
+      | 'hasAllowedEmailDomains'
+      | 'allowedEmailDomains'
+    > &
+      E2eFieldSingleValue)
+  | (E2ePickFieldMetadata<HomenoFieldBase, 'fieldType'> & E2eFieldSingleValue)
+  | (E2ePickFieldMetadata<ImageFieldBase, never> & E2eFieldFilepath)
+  | (E2ePickFieldMetadata<LongTextFieldBase, 'ValidationOptions'> &
+      E2eFieldSingleValue)
+  | (E2ePickFieldMetadata<
+      MobileFieldBase,
+      'isVerifiable' | 'allowIntlNumbers'
+    > &
+      E2eFieldSingleValue)
+  | (E2ePickFieldMetadata<NricFieldBase, never> & E2eFieldSingleValue)
+  | (E2ePickFieldMetadata<NumberFieldBase, 'ValidationOptions'> &
+      E2eFieldSingleValue)
+  | (E2ePickFieldMetadata<
+      RadioFieldBase,
+      'fieldOptions' | 'othersRadioButton'
+    > &
+      E2eFieldSingleValue)
+  | (E2ePickFieldMetadata<RatingFieldBase, 'ratingOptions'> &
+      E2eFieldSingleValue)
+  | E2ePickFieldMetadata<SectionFieldBase, never>
+  | (E2ePickFieldMetadata<
+      ShortTextFieldBase,
+      'ValidationOptions' | 'allowPrefill'
+    > &
+      E2eFieldSingleValue)
+  | E2ePickFieldMetadata<StatementFieldBase, never>
+  | (E2ePickFieldMetadata<
+      TableFieldBase,
+      'minimumRows' | 'addMoreRows' | 'columns' | 'maximumRows'
+    > &
+      E2eFieldTableValue)
+  | (E2ePickFieldMetadata<UenFieldBase, never> & E2eFieldSingleValue)
+  | (E2ePickFieldMetadata<YesNoFieldBase, never> & E2eFieldSingleValue)
+
+const allFieldInfo: E2eFieldMetadata[] = [
+  {
+    title: 'Attachment',
+    fieldType: BasicField.Attachment,
+    attachmentSize: AttachmentSize.OneMb,
+    val: 'test-att.txt',
+    path: '../files/att-folder-1/test-att.txt',
+  },
+  {
+    title: 'How did you hear about the event?',
+    fieldType: BasicField.Checkbox,
+    fieldOptions: ['School', 'Career Fairs / Talks', 'Online and Social Media'],
+    othersRadioButton: true,
+    ValidationOptions: { customMin: null, customMax: null },
+    validateByValue: false,
+    val: ['School', 'Mailing list'],
+  },
+  {
+    title: 'Birthday',
+    fieldType: BasicField.Date,
+    val: format(17, 'dd MMM yyyy'),
+    dateValidation: {
+      customMaxDate: null,
+      customMinDate: null,
+      selectedDateValidation: null,
+    },
+  },
+  {
+    title: 'Pi',
+    fieldType: BasicField.Decimal,
+    ValidationOptions: {
+      customMin: 3,
+      customMax: null,
+    },
+    validateByValue: true,
+    val: '3.1415926535',
+  },
+  {
+    title: 'Favourite Food',
+    fieldType: BasicField.Dropdown,
+    fieldOptions: ['Rice', 'Chocolate', 'Ice-Cream'],
+    val: 'Chocolate',
+  },
+  {
+    title: 'Personal Email',
+    fieldType: BasicField.Email,
+    isVerifiable: false,
+    autoReplyOptions: {
+      hasAutoReply: false,
+      autoReplyMessage: '',
+      autoReplySender: '',
+      autoReplySubject: '',
+      includeFormSummary: false,
+    },
+    hasAllowedEmailDomains: false,
+    allowedEmailDomains: [],
+    val: 'test@test.gov.sg',
+  },
+  {
+    title: 'Home Phone Number',
+    fieldType: BasicField.HomeNo,
+    val: '61234567',
+  },
+  // Hide for now, because it doesn't work unless we spin up aws local.
+  // {
+  //   title: 'Image',
+  //   fieldType: BasicField.Image,
+  //   path: '__tests__/e2e/files/logo.jpg',
+  //   description: 'Logo for an organization.',
+  // },
+  {
+    title: 'Your Life Story',
+    fieldType: BasicField.LongText,
+    ValidationOptions: {
+      selectedValidation: null,
+      customVal: null,
+    },
+    val: 'Vestibulum sed facilisis nibh, vel semper nisl. Phasellus dictum sem et ligula vulputate malesuada. Phasellus posuere luctus sapien eu molestie. In euismod vestibulum orci ac blandit. Suspendisse est arcu, vestibulum id viverra sed, tempus in ligula. Maecenas consequat pharetra lorem, ac vulputate neque sodales non. Mauris vel lacus ipsum.',
+  },
+  {
+    title: 'Mobile',
+    fieldType: BasicField.Mobile,
+    isVerifiable: false,
+    allowIntlNumbers: false,
+    val: '+6598889999',
+  },
+  {
+    title: 'NRIC Number',
+    fieldType: BasicField.Nric,
+    val: 'S9991334G',
+  },
+  {
+    title: 'Answer to life, the universe and everything?',
+    fieldType: BasicField.Number,
+    ValidationOptions: {
+      selectedValidation: null,
+      customVal: null,
+    },
+    val: '42',
+  },
+  {
+    title: 'Mother Tongue Language',
+    fieldType: BasicField.Radio,
+    fieldOptions: ['Chinese', 'Malay', 'Tamil'],
+    othersRadioButton: true,
+    val: 'Klingon',
+  },
+  {
+    title: 'Happiness Score',
+    fieldType: BasicField.Rating,
+    ratingOptions: {
+      steps: 5,
+      shape: RatingShape.Heart,
+    },
+    val: '3',
+  },
+  {
+    title: 'About you',
+    fieldType: BasicField.Section,
+  },
+  {
+    title: 'Name',
+    fieldType: BasicField.ShortText,
+    ValidationOptions: {
+      selectedValidation: null,
+      customVal: null,
+    },
+    val: 'Lorem Ipsum',
+  },
+  {
+    title: 'Paragraph',
+    fieldType: BasicField.Statement,
+    description:
+      'This is a paragraph with **markdown** and a url to https://open.gov.sg/.\n\n* Please type your answers here.\n* Please press submit once you are done.\n\nThank you!',
+  },
+  {
+    title: 'Family Members',
+    fieldType: BasicField.Table,
+    minimumRows: 2,
+    addMoreRows: false,
+    columns: [
+      {
+        title: 'Name',
+        required: true,
+        columnType: BasicField.ShortText,
+        ValidationOptions: {
+          customVal: null,
+          selectedValidation: null,
+        },
+      },
+      {
+        title: 'Gender',
+        required: true,
+        fieldOptions: ['Male', 'Female', 'Prefer not to say'],
+        columnType: BasicField.Dropdown,
+      },
+    ],
+    val: [
+      ['John', 'Male'],
+      ['Lisa', 'Female'],
+    ],
+  },
+  {
+    title: 'Some UEN here please',
+    fieldType: BasicField.Uen,
+    val: 'T09LL0001B',
+  },
+  {
+    title: 'Yes or No?',
+    fieldType: BasicField.YesNo,
+    val: 'Yes',
+  },
+]
+
+export const allFields = keyBy(allFieldInfo, 'fieldType')
