+++ __tests__/e2e/constants/links.ts
@@ -1,0 +1,5 @@
+export const ROOT_PAGE = `${process.env.APP_URL}`
+
+export const LOGIN_PAGE = `${ROOT_PAGE}/login`
+export const DASHBOARD_PAGE = `${ROOT_PAGE}/dashboard`
+export const ADMIN_FORM_PAGE_PREFIX = `${ROOT_PAGE}/admin/form`
